Adding the Language in the URL and initialize it with "en" as a default value that represents the English language for the startup of the website.
Replacing the old code for using the cooke in the log in, with a new code that takes the value of the language and use it to redirect to the log in page

Adding a RTL(Right to Left ) css file to support Arabic language.

Adding New folder named Resources that contains resourses files. There are two files: one with the name Language.resx contains the English language. The Language.ar.resx contains the Arabic Language translation.

Modifying all controllers to have the changingLanguageFunction that takes string language as a parameter which is sent when clicking the language buttons(English or Arabic)
and sending this language to Global.asx.cs  throgh cookie. The job of this Global class is to change the culture and UIculture depending on the value of the language variable that is sent to it so the website keeps reading from the targeted .resx resource file to show the pages in the desired language.

Calling the ChangingLanguageFunction in all action inside the controllers to allow the user to swich languages wherever he wants.

Replacing all string messages that appears for the user with new names that are written in the resource files with thier correspoding words for each language(projectName.ResourceFolder.ResourceFileName.TheNameOfTheStringInTheResourceFile)

Using the ViewContext.RouteData.Values["action"] or ViewContext.RouteData.Values["controller"] instead of specifying them in each changing language button, so it takes the action and the controller names from the where you are using the url.

Addding New Layouts (Details and Edit) and adding new parameter to changing language buttons (id=1) to be passed to some actions that must receives two paramater, so you can still swich languages inside them.

Changing the layouts of the views of these action. The ProgramDetails, EligibleList, and ClubForm now returns the views under DetailsLayout or EditLayout.