Adding new layouts with RTL CSS to support Arabic language.
Adding new views under the arabic layouts to show the website in RTL direction.

Modifying all controllers by adding if statements to each action in the controllers to select which view of which language to show depending on the language parameter.

Rearranging the views folder to seperate the English views from the arabic views.

Changing some mistakes in the arabic translation. Adding some missing words to be translated when using the arabic site.

Changing the names of the views to make it easier to distinguish between the english and arabic views

Adding two new actrions to the program controller for watermarking a certification, one for watermarking new certificate, and another is for watermarking an existing certificate in the project.

Adding some missing controller's names in some buttons.Manage, club, and program controllers.