Modifying the program controller to return both language views arabic and english.
Adding some new labels in .resx files, changing some mistakes in the arabic translation.

Modifying the Log in view to support multi language.

Correcting some labels in the CompletePrograms view

Correcting translation mistake in the Registration view

Adding mising a column in the StuPrograms view.

Adding the Completedprograms buttons in the navbar under student and clubCor layouts.

Correcting the layout of the student home.