Adding new theme to the website.
Adding new dynamic pages to the home which is for the visitors.

Adding some text to the home pages.

Changing the homeController to support the visitor view.

Modifying the program and club controllers to serve a new viewModel (Incomplete)

Adding some pictures to the Images file for the home page.

Adding new script under the scripts to give dynamic to the home page.

Adding new partial views to handle the new home page with the new style.

Changing the ViewStart to the new layout style.