Adding new cancel buttons in each form for  both arabic and english views.
Removing the unwanted actions for visiter with thier views in the club and program controllers.

Changing some string labels.

Seperating the program registration form into two views, Add view and Edit view for both ar and en.

Solving problem with PIDParameter layouts to handle the pass the right id to each language actionLink(ar, en).

Corrcting some of breadCrumbs for some views to give the right path.

Modifying the club controller by adding an if statments to determine the current action to provide the right breadCumbs for the view.