Installing bootbox package, JS files for showing alerts.
Adding the details modal to the arabic views.

Adding the bootbox.js and the alert.js to the BundleConfig.

Adding the navbar of the visitor home laytout to the student, clubcordinator, admin home layouts.

Fixing the menu of the visitor home layout where there was an error when resize the page to a small size.