Installing the bootbox alerts package in the project
Adding new confirmation messages to the buttons. Using the JavaScript file named "Alerts" that uses the bootbox alerts  for Delete, Leave, Approve, Reject buttons. Using the modal isnide the view to show the confirmation messages of the buttons that submits forms. Using the bootstrap alerts for showing the alerts that comes from the database with the help of the TempData and the ViewBag for sending these messages from the controllers to the views.

Modifying the StuProg to show only the registered programs for the student without the deleted programs.

Adding ViewBags for getting the title of the programs to show them in the views, such as the title of the view page, the BreadCrump.

Modifying the .resx to handle the confirmation messages in both languages.

Modifying the Datepicker. js to show the closest date.

Adding new script for the EndDate to reistrict the datepicker to show the date after the current date.

Changing some of the tables styles to show the data in better way. Also, setting different colors for the buttons.

Adding comments on the new changes.